webui: move model selector into ChatForm (idea by @allozaur)
